Fixed size_t conversion warnings. JIRA: CXX-5057
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@65025 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ include/db/bdb/bdb_bv_store.hpp
@@ -469,7 +469,7 @@ EBDB_ErrCode CBDB_BvStore<TBV>::FetchToBuffer(CBDB_FileCursor& cur)
     }
     catch (CBDB_ErrnoException& e) {
         if (e.IsBufferSmall()  ||  e.IsNoMem()) {
-            unsigned buf_size = LobSize();
+            size_t buf_size = LobSize();
             m_Buffer.resize(buf_size);
             void* p = &m_Buffer[0];
             err = cur.Fetch(CBDB_FileCursor::eDefault,
